seperate tests in each file and create more tests
