+++ polls/test/test_index.py
@@ -1,0 +1,66 @@
+"""TEST index in polls app."""
+import datetime
+from django.test import TestCase
+from django.utils import timezone
+from django.urls import reverse
+
+from polls.models import Question
+
+
+def create_question(question_text, days, closed):
+    """
+    Create a question with the given `question_text`.
+
+    And published the given number of `days` offset to now
+    (negative for questions published in the past, positive for questions that have yet to be published).
+    """
+    time = timezone.now() + datetime.timedelta(days=days)
+    closed = timezone.now() + datetime.timedelta(days=closed)
+    return Question.objects.create(question_text=question_text, pub_date=time, end_date=closed)
+
+
+class QuestionIndexViewTests(TestCase):
+    """Test the index page."""
+
+    def test_no_questions(self):
+        """If no questions exist, an appropriate message is displayed."""
+        response = self.client.get(reverse('polls:index'))
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerysetEqual(response.context['latest_question_list'], [])
+
+    def test_past_question(self):
+        """Questions with a pub_date in the past are displayed on the index page."""
+        create_question(question_text="Past question.", days=-30, closed=-20)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question.>']
+        )
+
+    def test_future_question(self):
+        """Questions with a pub_date in the future aren't displayed on the index page."""
+        create_question(question_text="Future question.", days=30, closed=20)
+        response = self.client.get(reverse('polls:index'))
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerysetEqual(response.context['latest_question_list'], [])
+
+    def test_future_question_and_past_question(self):
+        """Even if both past and future questions exist, only past questions are displayed."""
+        create_question(question_text="Past question.", days=-3, closed=-20)
+        create_question(question_text="Future question.", days=30, closed=50)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question.>']
+        )
+
+    def test_two_past_questions(self):
+        """The questions index page may display multiple questions."""
+        create_question(question_text="Past question 1.", days=-30, closed=-40)
+        create_question(question_text="Past question 2.", days=-5, closed=-7)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question 2.>', '<Question: Past question 1.>']
+        )
