+++ polls/test/test_polls_date.py
@@ -4,7 +4,7 @@ from django.test import TestCase
 from django.utils import timezone
 from django.urls import reverse
 
-from .models import Question
+from polls.models import Question
 
 
 def create_question(question_text, days, closed):
@@ -86,66 +86,3 @@ class QuestionModelTests(TestCase):
         self.assertIs(recent_question.is_published(), True)
 
 
-class QuestionIndexViewTests(TestCase):
-    """Test the index page."""
-
-    def test_no_questions(self):
-        """If no questions exist, an appropriate message is displayed."""
-        response = self.client.get(reverse('polls:index'))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerysetEqual(response.context['latest_question_list'], [])
-
-    def test_past_question(self):
-        """Questions with a pub_date in the past are displayed on the index page."""
-        create_question(question_text="Past question.", days=-30, closed=-20)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question.>']
-        )
-
-    def test_future_question(self):
-        """Questions with a pub_date in the future aren't displayed on the index page."""
-        create_question(question_text="Future question.", days=30, closed=20)
-        response = self.client.get(reverse('polls:index'))
-        self.assertContains(response, "No polls are available.")
-        self.assertQuerysetEqual(response.context['latest_question_list'], [])
-
-    def test_future_question_and_past_question(self):
-        """Even if both past and future questions exist, only past questions are displayed."""
-        create_question(question_text="Past question.", days=-3, closed=-20)
-        create_question(question_text="Future question.", days=30, closed=50)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question.>']
-        )
-
-    def test_two_past_questions(self):
-        """The questions index page may display multiple questions."""
-        create_question(question_text="Past question 1.", days=-30, closed=-40)
-        create_question(question_text="Past question 2.", days=-5, closed=-7)
-        response = self.client.get(reverse('polls:index'))
-        self.assertQuerysetEqual(
-            response.context['latest_question_list'],
-            ['<Question: Past question 2.>', '<Question: Past question 1.>']
-        )
-
-
-class QuestionDetailViewTests(TestCase):
-    """Test the detail page."""
-
-    def test_future_question(self):
-        """The detail view of a question with a pub_date in the future returns a 302 not found."""
-        future_question = create_question(question_text='Future question.', days=5, closed=10)
-        url = reverse('polls:detail', args=(future_question.id,))
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
-
-    def test_past_question(self):
-        """The detail view of a question with a pub_date in the past displays the question's text."""
-        past_question = create_question(question_text='Past Question.', days=-5, closed=-10)
-        url = reverse('polls:detail', args=(past_question.id,))
-        response = self.client.get(url)
-        self.assertEqual(response.status_code, 302)
